move hooks into cli
